Move config helpers from cmd/util to their own pkg
This stops inclusion of unneccessary dependencies when using Nitro from
other projects.
