+++ cmd/daserver/daserver.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/offchainlabs/nitro/arbnode"
 	"github.com/offchainlabs/nitro/cmd/genericconf"
-	"github.com/offchainlabs/nitro/cmd/util"
+	"github.com/offchainlabs/nitro/cmd/util/confighelpers"
 	"github.com/offchainlabs/nitro/das"
 )
 
@@ -91,17 +91,17 @@ func parseDAServer(args []string) (*DAServerConfig, error) {
 	das.DataAvailabilityConfigAddOptions("data-availability", f)
 	genericconf.ConfConfigAddOptions("conf", f)
 
-	k, err := util.BeginCommonParse(f, args)
+	k, err := confighelpers.BeginCommonParse(f, args)
 	if err != nil {
 		return nil, err
 	}
 
 	var serverConfig DAServerConfig
-	if err := util.EndCommonParse(k, &serverConfig); err != nil {
+	if err := confighelpers.EndCommonParse(k, &serverConfig); err != nil {
 		return nil, err
 	}
 	if serverConfig.ConfConfig.Dump {
-		err = util.DumpConfig(k, map[string]interface{}{
+		err = confighelpers.DumpConfig(k, map[string]interface{}{
 			"data-availability.key.priv-key": "",
 		})
 		if err != nil {
@@ -126,11 +126,11 @@ func startup() error {
 
 	serverConfig, err := parseDAServer(os.Args[1:])
 	if err != nil {
-		util.HandleError(err, printSampleUsage)
+		confighelpers.HandleError(err, printSampleUsage)
 		return nil
 	}
 	if !(serverConfig.EnableRPC || serverConfig.EnableREST) {
-		util.HandleError(nil, printSampleUsage)
+		confighelpers.HandleError(nil, printSampleUsage)
 		fmt.Printf("Please specify at least one of --enable-rest or --enable-rpc\n")
 		return nil
 	}
@@ -162,7 +162,7 @@ func startup() error {
 		return err
 	}
 
-	vcsRevision, vcsTime := util.GetVersion()
+	vcsRevision, vcsTime := confighelpers.GetVersion()
 	var rpcServer *http.Server
 	if serverConfig.EnableRPC {
 		log.Info("Starting HTTP-RPC server", "addr", serverConfig.RPCAddr, "port", serverConfig.RPCPort, "revision", vcsRevision, "vcs.time", vcsTime)
