+++ cmd/datool/datool.go
@@ -23,8 +23,9 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/offchainlabs/nitro/arbstate"
 	"github.com/offchainlabs/nitro/cmd/genericconf"
-
 	"github.com/offchainlabs/nitro/cmd/util"
+
+	"github.com/offchainlabs/nitro/cmd/util/confighelpers"
 	"github.com/offchainlabs/nitro/das"
 	"github.com/offchainlabs/nitro/das/dastree"
 	"github.com/offchainlabs/nitro/util/signature"
@@ -101,13 +102,13 @@ func parseClientStoreConfig(args []string) (*ClientStoreConfig, error) {
 	f.Duration("das-retention-period", 24*time.Hour, "The period which DASes are requested to retain the stored batches.")
 	genericconf.ConfConfigAddOptions("conf", f)
 
-	k, err := util.BeginCommonParse(f, args)
+	k, err := confighelpers.BeginCommonParse(f, args)
 	if err != nil {
 		return nil, err
 	}
 
 	var config ClientStoreConfig
-	if err := util.EndCommonParse(k, &config); err != nil {
+	if err := confighelpers.EndCommonParse(k, &config); err != nil {
 		return nil, err
 	}
 	return &config, nil
@@ -204,13 +205,13 @@ func parseRESTClientGetByHashConfig(args []string) (*RESTClientGetByHashConfig, 
 
 	genericconf.ConfConfigAddOptions("conf", f)
 
-	k, err := util.BeginCommonParse(f, args)
+	k, err := confighelpers.BeginCommonParse(f, args)
 	if err != nil {
 		return nil, err
 	}
 
 	var config RESTClientGetByHashConfig
-	if err := util.EndCommonParse(k, &config); err != nil {
+	if err := confighelpers.EndCommonParse(k, &config); err != nil {
 		return nil, err
 	}
 	return &config, nil
@@ -266,13 +267,13 @@ func parseKeyGenConfig(args []string) (*KeyGenConfig, error) {
 	f.Bool("wallet", false, "generate the ECDSA keypair in a wallet file")
 	genericconf.ConfConfigAddOptions("conf", f)
 
-	k, err := util.BeginCommonParse(f, args)
+	k, err := confighelpers.BeginCommonParse(f, args)
 	if err != nil {
 		return nil, err
 	}
 
 	var config KeyGenConfig
-	if err := util.EndCommonParse(k, &config); err != nil {
+	if err := confighelpers.EndCommonParse(k, &config); err != nil {
 		return nil, err
 	}
 	return &config, nil
