+++ cmd/nitro/nitro.go
@@ -40,6 +40,7 @@ import (
 	"github.com/offchainlabs/nitro/cmd/conf"
 	"github.com/offchainlabs/nitro/cmd/genericconf"
 	"github.com/offchainlabs/nitro/cmd/util"
+	"github.com/offchainlabs/nitro/cmd/util/confighelpers"
 	_ "github.com/offchainlabs/nitro/nodeInterface"
 	"github.com/offchainlabs/nitro/util/colors"
 	"github.com/offchainlabs/nitro/util/headerreader"
@@ -122,7 +123,7 @@ func main() {
 	args := os.Args[1:]
 	nodeConfig, l1Wallet, l2DevWallet, l1Client, l1ChainId, err := ParseNode(ctx, args)
 	if err != nil {
-		util.HandleError(err, printSampleUsage)
+		confighelpers.HandleError(err, printSampleUsage)
 
 		return
 	}
@@ -131,7 +132,7 @@ func main() {
 		panic(err)
 	}
 
-	vcsRevision, vcsTime := util.GetVersion()
+	vcsRevision, vcsTime := confighelpers.GetVersion()
 	log.Info("Running Arbitrum nitro node", "revision", vcsRevision, "vcs.time", vcsTime)
 
 	if nodeConfig.Node.Dangerous.NoL1Listener {
@@ -255,7 +256,7 @@ func main() {
 
 	chainDb, l2BlockChain, err := openInitializeChainDb(ctx, stack, nodeConfig, new(big.Int).SetUint64(nodeConfig.L2.ChainID), arbnode.DefaultCacheConfigFor(stack, &nodeConfig.Node.Caching))
 	if err != nil {
-		util.HandleError(err, printSampleUsage)
+		confighelpers.HandleError(err, printSampleUsage)
 		return
 	}
 
@@ -447,7 +448,7 @@ func ParseNode(ctx context.Context, args []string) (*NodeConfig, *genericconf.Wa
 
 	NodeConfigAddOptions(f)
 
-	k, err := util.BeginCommonParse(f, args)
+	k, err := confighelpers.BeginCommonParse(f, args)
 	if err != nil {
 		return nil, nil, nil, nil, nil, err
 	}
@@ -552,19 +553,19 @@ func ParseNode(ctx context.Context, args []string) (*NodeConfig, *genericconf.Wa
 		}
 	}
 
-	err = util.ApplyOverrides(f, k)
+	err = confighelpers.ApplyOverrides(f, k)
 	if err != nil {
 		return nil, nil, nil, nil, nil, err
 	}
 
 	var nodeConfig NodeConfig
-	if err := util.EndCommonParse(k, &nodeConfig); err != nil {
+	if err := confighelpers.EndCommonParse(k, &nodeConfig); err != nil {
 		return nil, nil, nil, nil, nil, err
 	}
 
 	// Don't print wallet passwords
 	if nodeConfig.Conf.Dump {
-		err = util.DumpConfig(k, map[string]interface{}{
+		err = confighelpers.DumpConfig(k, map[string]interface{}{
 			"l1.wallet.password":        "",
 			"l1.wallet.private-key":     "",
 			"l2.dev-wallet.password":    "",
