+++ cmd/util/confighelpers/configuration.go
@@ -1,7 +1,7 @@
 // Copyright 2021-2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
-package util
+package confighelpers
 
 import (
 	"fmt"
